requirements.txt
# Web Crawler Requirements
seleniumbase>=4.45.0
beautifulsoup4>=4.12.0
requests>=2.32.0
pymongo>=4.6.0
schedule>=1.2.0
flask>=3.1.0
pandas>=2.2.0
python-dateutil>=2.8.2
PyPDF2>=3.0.0
python-docx>=1.1.0
feedparser>=6.0.10
python-dotenv>=1.0.0

# AI Dependencies (Lightweight local models)
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
